Cluster interface removed
This commit removes Cluster interface and all things related to clusters in the public API.

Until now the Cluster interface provided a possibility to change syntax tree manually, however in the current design such capabilities are almost useless as the Analyzer does not expose them.

Another purpose of the Cluster was public representation of the MutableUnit inner caches, which in fact is an implementation detail, and should have been encapsulated from the beginning. Having this interface in the public API complicates the entire design of the project and imposes unnecessary complexity.

The drawback of this change is that an API user is no longer able to restructure syntax tree manually. However the commit kept inner algorithms intact. As such, if needed, this feature could be returned back in the future versions in more concise way.

Changes in the "arena" module:

 - Entry object simplified. From now on it only stores index and version.
 - The "Primary" variant is no longer expressed within Entry as the primary entries removed from the public API design alongside with Clusters.
 - Repository renamed to Repo. It's interface and implementation adjusted to mimic necessary commit-wide changes.
 - In particular Repo version starts with 1.
 - As such Entries with version 0 encode indexed data outside of repositories.
 - Entries with index and version set to max values express Nil entries.
 - Sequence object removed. An API encouraged to use just normal std Vec objects instead.

Changes in the "syntax" module:

 - Cluster interface removed.
 - NodeRef's and ErrorRef's "cluster_entry" field removed.
 - All cluster operations related to NodeRef and ErrorRef removed.
 - All operations in the SyntaxTree trait related to clusters removed.
 - Node and Error objects access function added to the SyntaxTree trait.
 - SyntaxTree::NodeIterator and SyntaxTree::ErrorIterator associated types added.
 - These traits assume to expose iterators over the NodeRefs and ErrorRefs respectively and are implementation dependent.
 - In general nodes and errors iteration became more trivial procedure as both types of objects from now on stored in flat collections under the hood of the Mutable and Immutable units.
 - SyntaxBuffer renamed to ImmutableSyntaxTree.

Changes in the "units" module:
 - MutableUnit internal implementation no longer depends on the previous Cluster interface.
 - The internal algorithm kept almost intact, and the unit still operates with node caches under the hood. However the nodes and the eroors are stored aside from the storage tree in flat Repo collections.

Among other things these changes fix a problem with the root's leftmost nodes caching. Previously they have been treated as secondary nodes regardless of their caching capabilities. From now on they will be cached properly.

Also this set of changes slightly improves MutableUnit performance.