Release 2.0 preparations
Final code changes before the release:

 - No-std support removed.
 - Inner refactoring in the panic reports functions.
 - mem module tests added.
 - Oudated tests removed.
 - ParseSegment renamed to ParseNodeChild.
 - ParseError renamed to SyntaxError.
 - Node::Error assosciative type removed. The syntax error type from now on is no longer customizable.
 - NoSyntax renamed to VoidSyntax.
 - Node macro "index" macro attribute renamed to "denote".
 - Watch interface renamed to Watcher.
 - Table::drain function behavior fixed to properly mimic HashMap::drain when the TableDrain iterator dropped without full consumption.
 - Latch renamed to Trigger. Latch::get function removed. Latch::get_relaxed function renamed to "is_active". Latch::set renamed to "activate".
 - All debug and display functions renamed to just "display", and return "impl Debug + Display" instead of concrete types.
 - NIL_SITE_REF static introduced.
 - ImmutableSyntaxTree id cleanup bug fixed (subid).
 - Analyzer's task manager refactored. From now on task acquiring is based on the explicit priority (TaskPriority numeric type) instead of the task type. All corresponding Analyzer functions fixed accordingly.
 - Analyzer's Handle object replaced with customizable TaskHandle trait.
 - TriggerHandle is a default implementation of the TaskHandle backed by a single Trigger (former Latch). The API user supposed to implement more complex triggering logic manually depending on the application domain.
 - AnalysisError::DuplicateHandle variane removed, because handles uniqueness is no longer a requirement.
 - format::Priority enum renamed to AnnotationPriority.
 - Token macro's "constructor" macro attribute from now on accepts arbitrary expressions.
 - Fixed a bug in the MutableUnit lexer.
 - Warnings and todos cleanup.
 - Crates MSRV raised to "1.79".
 - Crates categories updated, versions upgraded to "2.0.0".

Release preparations:

 - Old API documentation replaced with the new one.
 - Lady Deirdre Examples implemented.
 - User Guide book added.
 - Benchmark tests improved.
 - All readme's fixed in accordance to changes.
 - End-User License Agreement replaced with the new General License Agreement.

There are a few more things to address before the release, including the book deployment, and also fixing various links across the files. The changes will be addresses in the addendum commits.

The new License Agreement from now on will govern Lady Deirdre public distribution and use. The core concepts behind the document stay intact: the users can still use my project for personal, non-commercial and the limited commercial use without the need to pay me royalties or fees, and the users required to assign me changes and contributions to Lady Deirdre. However, the new document clarifies several aspects and legal nuances more accurately.

The major change in the document is the replacement of the requirement to pay me roylties for use of the project in commercial products that exceed free commercial limits. The new agreement obligates the user to acquire Unrestricted Commercial License upon reaching of the commercial limits set forth in the agreement. This license is acquired by the user independently from the agreement, and allows continuation of use of Lady Deirdre in one of the user's commercial product. I plan to sell commercial licenses separately using dedicated platforms.

To sum up, the new agreement effectively replaces the model of royalties with the model of the flat fees through the commercial licences purchasing.